LC-single-number-iii [Runtime: N/A; Memory: N/A]

+++ array/single-number-iii.cpp
@@ -9,7 +9,7 @@ public:
             ans ^= n;
         }
 
-        int mask = ans & (-ans);
+        long long mask = ans & (-ans);
 
         int a = 0, b = 0;
